requirements.txt
celery
Django
djangorestframework
# django-celery-email
psycopg2>=2.8
